# api/openapi.yaml
# Code generated by go run ./cmd/openapi generate; DO NOT EDIT.
openapi: 3.0.3
info:
  title: "Keiyaku-Go CMS API"
  version: "1.0.0"
servers:
  - url: "/api/v1"
paths:
  /articles:
    get:
      tags:
        - "Article"
      summary: "List published articles"
      description: "List published articles with pagination."
      parameters:
        - name: "page"
          in: query
          required: false
          description: "Page number"
          schema:
            type: integer
            format: int32
        - name: "page_size"
          in: query
          required: false
          description: "Page size"
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/ArticleListResponse"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
    post:
      tags:
        - "Article"
      summary: "Create article"
      description: "Create an article for the authenticated user."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: "Article payload"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/ArticleResponse"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
  /articles/{id}:
    get:
      tags:
        - "Article"
      summary: "Get published article"
      description: "Get a published article by ID."
      parameters:
        - name: "id"
          in: path
          required: true
          description: "Article ID"
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/ArticleResponse"
        "400":
          description: "Invalid article ID"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "404":
          description: "Article not found"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
  /auth/login:
    post:
      tags:
        - "Auth"
      summary: "Login user"
      description: "Authenticate a user and return access and refresh tokens."
      requestBody:
        required: true
        description: "Login payload"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/AuthResponse"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "422":
          description: "Invalid credential"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
  /auth/register:
    post:
      tags:
        - "Auth"
      summary: "Register user"
      description: "Register a user and return access and refresh tokens."
      requestBody:
        required: true
        description: "Register payload"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/AuthResponse"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "409":
          description: "Conflict"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
  /users/me:
    get:
      tags:
        - "User"
      summary: "Current user profile"
      description: "Get the authenticated user's profile."
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
                  data:
                    $ref: "#/components/schemas/UserResponse"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - msg
                properties:
                  code:
                    type: integer
                    format: int32
                  msg:
                    type: string
components:
  schemas:
    ArticleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArticleResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    ArticleResponse:
      type: object
      properties:
        author_id:
          type: integer
          format: int64
        category_id:
          type: integer
          format: int64
        content:
          type: string
        created_at:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        published_at:
          type: string
          format: date-time
          nullable: true
        slug:
          type: string
        status:
          type: string
        summary:
          type: string
        tags:
          type: array
          items:
            type: string
        title:
          type: string
        updated_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
    CreateArticleRequest:
      type: object
      required:
        - content
        - slug
        - title
      properties:
        category_id:
          type: integer
          format: int64
        content:
          type: string
        publish:
          type: boolean
        slug:
          type: string
          maxLength: 180
        summary:
          type: string
          maxLength: 512
        tags:
          type: array
          items:
            type: string
        title:
          type: string
          maxLength: 160
    LoginRequest:
      type: object
      required:
        - password
        - username
      properties:
        password:
          type: string
        username:
          type: string
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        display_name:
          type: string
          maxLength: 64
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        username:
          type: string
          minLength: 3
          maxLength: 32
    UserResponse:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        display_name:
          type: string
        email:
          type: string
        id:
          type: integer
          format: int64
        roles:
          type: array
          items:
            type: string
        status:
          type: string
        updated_at:
          type: string
          format: date-time
        username:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT